test(moon): isolate registry search configuration
