+++ crates/moon/tests/test_cases/moon_commands/mod.rs
@@ -1,7 +1,5 @@
 use super::*;
 
-use std::io::{Read, Write};
-
 fn moonx_bin(bin_dir: &tempfile::TempDir) -> std::path::PathBuf {
     let moonx = bin_dir
         .path()
@@ -104,64 +102,6 @@ fn test_moon_help() {
                       Unstable flags to MoonBuild [env: MOON_UNSTABLE=] [default: ]
         "#]],
     );
-}
-
-#[test]
-fn test_moon_search_uses_configured_registry() {
-    let listener = std::net::TcpListener::bind("127.0.0.1:0").unwrap();
-    listener.set_nonblocking(true).unwrap();
-    let registry = format!("http://{}", listener.local_addr().unwrap());
-    let server = std::thread::spawn(move || {
-        let deadline = std::time::Instant::now() + std::time::Duration::from_secs(5);
-        let (mut stream, _) = loop {
-            match listener.accept() {
-                Ok(connection) => break connection,
-                Err(error) if error.kind() == std::io::ErrorKind::WouldBlock => {
-                    assert!(
-                        std::time::Instant::now() < deadline,
-                        "moon search did not contact the configured registry"
-                    );
-                    std::thread::sleep(std::time::Duration::from_millis(5));
-                }
-                Err(error) => panic!("registry test server failed: {error}"),
-            }
-        };
-
-        let mut request = Vec::new();
-        let mut buffer = [0; 1024];
-        while !request.windows(4).any(|window| window == b"\r\n\r\n") {
-            let read = stream.read(&mut buffer).unwrap();
-            assert_ne!(read, 0, "request ended before its headers");
-            request.extend_from_slice(&buffer[..read]);
-        }
-        assert!(
-            request.starts_with(b"GET /api/v0/search?kw=json+query&limit=2 HTTP/1.1\r\n"),
-            "unexpected registry request: {}",
-            String::from_utf8_lossy(&request)
-        );
-
-        let body = br#"[{"name":"example/json","version":"1.2.3","description":"JSON query\ntools\r\u001b[31mwith color\u001b[0m\tand spaces"},{"name":"example/no-description","version":"0.4.0"}]"#;
-        write!(
-            stream,
-            "HTTP/1.1 200 OK\r\nContent-Type: application/json\r\nContent-Length: {}\r\nConnection: close\r\n\r\n",
-            body.len()
-        )
-        .unwrap();
-        stream.write_all(body).unwrap();
-    });
-
-    let dir = TestDir::new_empty();
-    moon_cmd(&dir)
-        .env("MOONCAKES_REGISTRY", registry)
-        .env("NO_PROXY", "*")
-        .args(["search", "json query", "--limit", "2"])
-        .assert()
-        .success()
-        .stdout_eq(
-            "example/json@1.2.3: JSON query tools with color and spaces\nexample/no-description@0.4.0\n",
-        )
-        .stderr_eq("");
-    server.join().unwrap();
 }
 
 #[test]
